Major update

+++ Linux/SSH.sh
@@ -4,6 +4,7 @@
 # generating a passphrase (if needed) then a SSH key pair with a send to the remote server
 # ssh-keygen [-q] [-a rounds] [-b bits] [-C comment] [-f output_keyfile] [-m format] [-N new_passphrase] [-O option] [-t dsa | ecdsa | ecdsa-sk | ed25519 | ed25519-sk | rsa]
 
+
 #-----------------------#
 #	Functions	#
 #-----------------------#
@@ -24,29 +25,19 @@ CreatePassphrase()
 {
 	#	Creating a secure passphrase if needed
 	ConfirmChoice "Generate a passphrase with 128 bits of entropy ?" && passwd=`dd if=/dev/urandom bs=16 count=1 2>/dev/null | base64 | sed 's/=//g'`
-	echo $passwd > $HOME/passph.txt
-
 	ConfirmChoice "Generate a passphrase with 256 bits of entropy ?" && passwd=`dd if=/dev/urandom bs=32 count=1 2>/dev/null | sha256sum -b | sed 's/ .*//'`
-	echo $passwd > $HOME/passph.txt
 }
 
 #-------------------#
 #	Start	    #
 #-------------------#
 
-if	[ ! -d .ssh ] ; then
+if	[ ! -d ~/.ssh ] ; then
 	mkdir ~/.ssh
 fi
 
 #	Using an existing passphrase
-ConfirmChoice "Do you have a passphrase to use" && read -s passwd || CreatePassphrase
-
-if	[ ! -z $passwd ] ; then
-	chmod 400 $HOME/passph.txt && echo "Your passphrase is located in $HOME/passph.txt"
-fi
-
-##	Generation of a public/private key pair
-#	Encryption algorithms
+ConfirmChoice "Do you want to use a passphrase" && read -sp 'Your passphrase : ' passwd || CreatePassphrase
 
 while [ -z $keyname ] ; do
 	echo "Note : If the key name is already existing, you can choose to overwrite it or not"
@@ -54,11 +45,21 @@ while [ -z $keyname ] ; do
 	read -p 'Give your ssh key a name : ' keyname
 done
 
-echo "One of the encryption algorithms is recommended"
+------------------------------
+passph="passph-$keyname.txt"
+------------------------------
+
+##	Generation of a public/private key pair
+#	Encryption algorithms
+
+echo "One of these encryption algorithms is recommended"
 echo "RSA | ED25519" && sleep 2
 
 if 	[ ! -z $passwd ] ; then
-	ConfirmChoice "RSA , Proven and recommended with a key size of 4096 bits. Compatible everywhere" && ssh-keygen -a 100 -b 4096 -N "$passwd" -f ~/.ssh/$keyname # by default so no need for "-t"
+	echo $passwd > $HOME/$passph && chmod 400 $HOME/$passph ; echo "Your passphrase is located in $HOME/$passph"
+
+	# by default so no need for "-t"
+	ConfirmChoice "RSA , Proven and recommended with a key size of 4096 bits. Compatible everywhere" && ssh-keygen -a 100 -b 4096 -N "$passwd" -f ~/.ssh/$keyname
 
 	ConfirmChoice "ED25519 , The latest and greatest in terms of safety and performance" && ssh-keygen -a 100 -f ~/.ssh/$keyname -N "$passwd" -t ed25519
 
@@ -69,29 +70,29 @@ else
 	ConfirmChoice "ED25519 , The latest and greatest in terms of safety and performance" && ssh-keygen -a 100 -f ~/.ssh/$keyname -t ed25519
 fi
 
-read -p 'Type the address of the remote server in this form: (example: 192.168.1.2) ' address
+if	[ -f ~/.ssh/$keyname ] ; then
+	read -p 'Type the address of the remote server in this form: (example: 192.168.1.2) ' address
+	read -p 'Type in the remote server login - this is the server user : ' login
 
-read -p 'Type in the remote server login - this is the server user : ' login
+	#	Checking the data entered by the client
+	echo "your login and IP address are: $login@$address"
 
-#	Checking the data entered by the client
-echo "your login and IP address are: $login@$address"
+	#	Send the public key to the desired server
+	ssh-copy-id -i ~/.ssh/$keyname.pub $login@$address
 
-#	Send the public key to the desired server
-ssh-copy-id -i ~/.ssh/$keyname.pub $login@$address
+	#	Make SSH connection easier
+	echo "Creating ~/.ssh/config"
+	echo "Usage ssh <customised hostname>"
 
-#	Make SSH connection easier
-echo "Creating ~/.ssh/config"
-echo "Usage ssh <customised hostname>"
+	read -p 'Name your machine : ' hostname
 
-read -p 'Name your machine : ' hostname
-
-tee -a ~/.ssh/config << END
-Host $hostname
-	HostName $address
-	User $login
-	IdentityFile ~/.ssh/$keyname
-END
-
-#	Final step
-ConfirmChoice "Do you want to back up your public key to your $HOME ?" && cp -p ~/.ssh/$keyname.pub $HOME/$keyname.pub
-ConfirmChoice "Then remove it ?" && rm ~/.ssh/$keyname.pub
+	tee -a ~/.ssh/config << END
+	Host $hostname
+		HostName $address
+		User $login
+		IdentityFile ~/.ssh/$keyname
+	END
+else
+	echo 'No key pair created'
+	exit 3
+fi
